Release: Supervisor Routes for User Viewing
- Remove: User viewing functions from admin-controller
- Update: Stop using payloadUserId in getUser() in user-controller
- Update: Use convertedUserId from params in user-controller
- Update: Admin-routes use user viewing functions from user-controller
- Add: Supervisor-routes  use user viewing functions from user-controller
