+++ src/routes/supervisor-routes.js
@@ -1,5 +1,4 @@
 const express = require('express');
-const adminController = require('../controllers/admin-controller');
 const authMiddleware = require('../middleware/authentication-middleware');
 const authorizationMiddleware = require('../middleware/authorization-middleware');
 const methodNotAllowedMiddleware = require('../middleware/method-not-allowed-middleware');
@@ -8,20 +7,16 @@ const userController = require('../controllers/user-controller');
 const router = express.Router();
 
 router.use(authMiddleware);
-router.use(authorizationMiddleware('admin'));
+router.use(authorizationMiddleware('supervisor'));
 
 router
   .route('/users')
   .get(userController.getAllUsers)
-  .post(adminController.registerUsers)
   .all(methodNotAllowedMiddleware);
 
 router
   .route('/users/:userId')
   .get(userController.getUser)
-  .put(adminController.updateUser)
-  .patch(adminController.updateUser)
-  .delete(adminController.deleteUser)
   .all(methodNotAllowedMiddleware);
 
 module.exports = router;
